requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
jinja2==3.1.6
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.16
pypdf==5.4.0
docx2txt==0.8
python-dotenv==1.1.0
openai==1.76.0
scikit-learn==1.8.0
sentence-transformers==3.4.1
